Phase 3: zone→plan nomenclature cleanup (booking-view layer)
PLAN_officemap.md Phase 3. Mechanical, grep-driven rename of the booking-view
layer so 'zone' means only access-control (zones.py/zones.html/zoneAssign —
correctly named, untouched). No logic change. Domain terms (Zone, zone_type,
zone_group, zoneGroups, ZONE_ROLE_*, zone_show_* prefs, isZoneAdmin/Viewer)
are unchanged.

Renames:
- js/views/zone.js -> js/views/plan.js (collapsed the 5-line plan.js stub that
  imported zone.js; the booking view now lives in plan.js). zone.js deleted.
- js/views/modules/zoneuserdata.js -> planuserdata.js; class ZoneUserData ->
  PlanUserData (bookas.js import updated).
- js/views/css/zone/ -> js/views/css/plan/ (import updated).
- warp/xhr/zone.py -> warp/xhr/plan.py; Blueprint('zone') -> ('plan'),
  url_prefix 'zone' -> 'plan'; xhr/__init__.py import + register updated.
  (warp/xhr/zones.py — plural access-control CRUD — untouched.)
- plan.html DOM ids/classes: zonemap->planmap, zone_sidepanel->plan_sidepanel,
  zonemap_help_modal->planmap_help_modal, zone_container/zone_map/
  zone_action_btn/zone_datelist/zone_timeslider/zone_datetime_container->
  plan_*; url_for('xhr.zone.*') -> xhr.plan.*; warpGlobals keys zoneApply/
  zoneAutoBook/zoneGetUsers/zonePreviewPrefs -> plan*; view.py emits planPreviewPrefs.
- bookings.html + bookings.js: zoneApply -> planApply, xhr.zone.apply -> xhr.plan.apply.
- base.js: prefsSaved event payload zonePreviewPrefs -> planPreviewPrefs.
- style.css: booking-view .zone_* styling -> .plan_*; .zone_modify_* (editor)
  and .zone_action_btn (shared with the access-control assign modals, which
  keep that class) are left intact — the base margin rule now covers both
  .zone_action_btn and .plan_action_btn.
- e2e: booking-view selectors renamed (#zonemap->#planmap, .zone_action_btn->
  .plan_action_btn, etc.); the editor e2e (#zone_map_container) and access-control
  assign-modal references untouched. Also fixed a Phase-2 latent break:
  #zonemap > img (img removed) -> #planmap .OMBackground for the dark-filter test.

	full e2e: 293 passed.
