+++ e2e/tests/zone-admin/book-as.spec.ts
@@ -65,7 +65,7 @@ test.describe('booking as another user', () => {
     await expect(page.locator('#action_modal')).toHaveClass(/open/);
     await Promise.all([
       page.waitForResponse(r => r.url().includes('/xhr/zone/apply') && r.status() === 200),
-      page.locator('.zone_action_btn[data-action="book"]').click(),
+      page.locator('.plan_action_btn[data-action="book"]').click(),
     ]);
     await page.waitForLoadState('networkidle');
     await page.waitForTimeout(200);
