+++ js/views/modules/bookas.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import ZoneUserData from "./zoneuserdata";
+import PlanUserData from "./planuserdata";
 
 export default function BookAs() {
 
@@ -90,7 +90,7 @@ BookAs.prototype._onBlur = function(e) {
 
 BookAs.prototype._onAutocomplete = function(el,sel) {
 
-    var login = ZoneUserData.makeUserStrRev(sel);
+    var login = PlanUserData.makeUserStrRev(sel);
     this._setSelectedLogin(login);
 
     this._callChangeListeners();
@@ -156,6 +156,6 @@ BookAs.prototype._init =  function(zoneUserData) {
     }
 }
 
-ZoneUserData.getInstance().on('load',function(zoneUserData) {
+PlanUserData.getInstance().on('load',function(zoneUserData) {
     BookAs.getInstance()._init(zoneUserData);
     });
