+++ warp/templates/plan.html
@@ -4,21 +4,21 @@
     {{ super() }}
 
     <script>
-        window.warpGlobals.URLs['zoneApply'] = "{{ url_for('xhr.zone.apply') }}";
-        window.warpGlobals.URLs['getSeat'] = "{{ url_for('xhr.zone.getSeats',pid=pid) }}";
+        window.warpGlobals.URLs['planApply'] = "{{ url_for('xhr.plan.apply') }}";
+        window.warpGlobals.URLs['getSeat'] = "{{ url_for('xhr.plan.getSeats',pid=pid) }}";
         window.warpGlobals.URLs['seatSprite'] = "{{ url_for('static',filename='images/seat_icons.svg') }}";
         window.warpGlobals.URLs['planImage'] = "{{ url_for('view.planImage', pid=pid) }}";
         window.warpGlobals.darkFilter = {{ dark_filter | tojson }};
 
         window.warpGlobals['defaultSelectedDates'] = {{ defaultSelectedDates | tojson }};
         window.warpGlobals.today = {{ today | int }};
-        window.warpGlobals['zonePreviewPrefs'] = {{ zonePreviewPrefs | tojson }};
+        window.warpGlobals['planPreviewPrefs'] = {{ planPreviewPrefs | tojson }};
 
-        window.warpGlobals.URLs['zoneAutoBook'] = "{{ url_for('xhr.zone.autoBook', pid=pid) }}";
+        window.warpGlobals.URLs['planAutoBook'] = "{{ url_for('xhr.plan.autoBook', pid=pid) }}";
 
         {% if isZoneAdmin %}
         window.warpGlobals.isZoneAdmin = true;
-        window.warpGlobals.URLs['zoneGetUsers'] = "{{ url_for('xhr.zone.getUsers',pid=pid)}}";
+        window.warpGlobals.URLs['planGetUsers'] = "{{ url_for('xhr.plan.getUsers',pid=pid)}}";
         window.warpGlobals.daysInAdvance = {{ ((config['WEEKS_IN_ADVANCE'] + 1) * 7) | int }};
         {% elif isZoneViewer %}
         window.warpGlobals.isZoneViewer = true;
@@ -44,13 +44,13 @@
 
                 <div class="action_modal_buttons">
                     {% if isZoneAdmin %}
-                    <a href="#!" class="modal-close waves-effect btn-flat zone_action_btn TR" data-action="assign-modal">btn.Assign</a>
-                    <a href="#!" class="modal-close waves-effect btn-flat zone_action_btn TR" data-action="enable">btn.Enable</a>
-                    <a href="#!" class="modal-close waves-effect btn-flat zone_action_btn TR" data-action="disable">btn.Disable</a>
+                    <a href="#!" class="modal-close waves-effect btn-flat plan_action_btn TR" data-action="assign-modal">btn.Assign</a>
+                    <a href="#!" class="modal-close waves-effect btn-flat plan_action_btn TR" data-action="enable">btn.Enable</a>
+                    <a href="#!" class="modal-close waves-effect btn-flat plan_action_btn TR" data-action="disable">btn.Disable</a>
                     {% endif %}
-                    <a href="#!" class="modal-close waves-effect waves-light btn warp-btn-danger zone_action_btn TR" data-action="delete">btn.Remove</a>
-                    <a href="#!" class="modal-close waves-effect waves-light btn warp-btn-primary zone_action_btn TR" data-action="update">btn.Update</a>
-                    <a href="#!" class="modal-close waves-effect waves-light btn warp-btn-primary zone_action_btn TR" data-action="book">btn.Book</a>
+                    <a href="#!" class="modal-close waves-effect waves-light btn warp-btn-danger plan_action_btn TR" data-action="delete">btn.Remove</a>
+                    <a href="#!" class="modal-close waves-effect waves-light btn warp-btn-primary plan_action_btn TR" data-action="update">btn.Update</a>
+                    <a href="#!" class="modal-close waves-effect waves-light btn warp-btn-primary plan_action_btn TR" data-action="book">btn.Book</a>
                 </div>
             </div>
         </div>
@@ -69,13 +69,13 @@
             <ul id="assigned_seat_list" class="collection assigned_seat_list"></ul>
         </div>
         <div class="modal-footer">
-            <a href="#!" class="modal-close waves-effect waves-light btn zone_action_btn TR" data-action="assign">btn.Set</a>
+            <a href="#!" class="modal-close waves-effect waves-light btn plan_action_btn TR" data-action="assign">btn.Set</a>
             <a href="#!" class="modal-close waves-effect waves-light btn-flat TR">btn.Cancel</a>
         </div>
     </dialog>
     {% endif %}
 
-    <dialog id="zonemap_help_modal" class="modal modal-fixed-footer warp-form-modal">
+    <dialog id="planmap_help_modal" class="modal modal-fixed-footer warp-form-modal">
         <div class="modal-content">
             <table>
                 <tr>
@@ -128,25 +128,25 @@
 {% block content %}
     {{ super() }}
 
-    <div class="zone_container">
+    <div class="plan_container">
 
-        <div id="zone_sidepanel">
+        <div id="plan_sidepanel">
 
-            <div class="zone_sidepanel_topcontainer">
+            <div class="plan_sidepanel_topcontainer">
 
                 <div class="input-field book-as_container">
                     <input type="text" id="book-as" class="autocomplete book-as_input" autocomplete="off">
                     <label for="book-as" class="active TR">Book as</label>
                 </div>
 
-                <div class="zone_sidepanel_close sidenav-close" data-target="zone_sidepanel">
+                <div class="plan_sidepanel_close sidenav-close" data-target="plan_sidepanel">
                     <i class="material-icons">close</i>
                 </div>
             </div>
 
-            <div class="zone_datetime_container">
+            <div class="plan_datetime_container">
 
-                <div class="zone_datelist">
+                <div class="plan_datelist">
                     {% for w in nextWeek %}
                     <p>
                         <label>
@@ -159,7 +159,7 @@
                     {% endfor %}
                 </div>
 
-                <div class="zone_timeslider">
+                <div class="plan_timeslider">
                     <div id="timeslider-min" class="timeslider-hint">12:00</div>
                     <div id="timeslider" class="timeslider" data-min="{{ config['BOOK_OPEN'] }}" data-max="{{ config['BOOK_CLOSE'] }}"></div>
                     <div id="timeslider-max" class="timeslider-hint">12:00</div>
@@ -169,12 +169,12 @@
         </div>
 
 
-        <div class="zone_map" id="zonemap">
-            <div class="zonemap_help">
+        <div class="plan_map" id="planmap">
+            <div class="planmap_help">
                 {% include 'icons/help_icon.html' %}
             </div>
         </div>
-        <div class="zonemap_datetime_trigger sidenav-trigger" data-target="zone_sidepanel">
+        <div class="planmap_datetime_trigger sidenav-trigger" data-target="plan_sidepanel">
             {% include 'icons/schedule_icon_side.html' %}
         </div>
