+++ warp/xhr/__init__.py
@@ -1,7 +1,7 @@
 import flask
 
 from . import bookings
-from . import zone
+from . import plan
 from . import users
 from . import groups
 from . import zones
@@ -12,7 +12,7 @@ from . import calendar
 bp = flask.Blueprint('xhr', __name__)
 
 bp.register_blueprint(bookings.bp)
-bp.register_blueprint(zone.bp)
+bp.register_blueprint(plan.bp)
 bp.register_blueprint(users.bp)
 bp.register_blueprint(groups.bp)
 bp.register_blueprint(zones.bp)
